+++ cookbook/gravmag_basin2d_triangular.py
@@ -2,21 +2,15 @@
 GravMag: Simple gravity inversion for the relief of a 2D triangular basin
 """
 import numpy
-from fatiando import logger, utils, mesher, gravmag, inversion
+from fatiando import utils, mesher, gravmag, inversion
 from fatiando.vis import mpl
 
-log = logger.get()
-log.info(logger.header())
-log.info(__doc__)
-
-log.info("Generating synthetic data")
 verts = [(10000, 1.), (90000, 1.), (80000, 5000)]
 model = mesher.Polygon(verts, {'density':-100})
 xp = numpy.arange(0., 100000., 1000.)
 zp = numpy.zeros_like(xp)
 gz = utils.contaminate(gravmag.talwani.gz(xp, zp, [model]), 1)
 
-log.info("Preparing for the inversion")
 solver = inversion.gradient.levmarq(initial=(10000, 1000))
 estimate, residuals = gravmag.basin2d.triangular(xp, zp, gz, verts[0:2], -100,
     solver)