+++ cookbook/gravmag_imaging_sandwich2.py
@@ -2,12 +2,8 @@
 GravMag: 3D imaging using the sandwich model method on synthetic gravity data
 (more complex model)
 """
-from fatiando import logger, gridder, mesher, gravmag
+from fatiando import gridder, mesher, gravmag
 from fatiando.vis import mpl, myv
-
-log = logger.get()
-log.info(logger.header())
-log.info(__doc__)
 
 # Make some synthetic gravity data from a simple prism model
 prisms = [mesher.Prism(-4000,0,-4000,-2000,2000,5000,{'density':1200}),